Cleaned up output, moved time approximations to funcs.
Also added an approximation function which takes the next
numInstances+2 (usually 7) and takes the average of the difference
between the last 5 (default). This is a shitty way to estimate the
average interval, since it's easily avoidable.

Cleaning up the next interval output. It now says seconds where it
should.

A thing to fix: it usually says it's executing now but never does, for
whatever reason.