bot: delete bot action/notice replying
These weren't used anywhere at all... and replying by either of those
methods is actually kinda silly!

They should be opt-in if they exist. We can add em back if we need to.

Preferably they would be added on as specializations of the reply
function, not inferred magically by the type of message that triggered
it.