mastodon: initial module commit
This provides URL handling for ActivityPub Note and Person records;
namely, links to users and toots on Mastodon instances post v1.6.0.

A list of instances to automatically handle is included as
default_instance.json. A future commit will add instance discovery as
described in the first comment of index.js.

The module handles robots.txt files that *explicitly* exclude this bot.
info.md is referenced in the User-Agent string so that instance
operators know what the requests are coming from and what requests it
can make.