url: consistently use 'new URL'
'new URL' and 'url.parse' handle different sets of urls.

Using them inconsistently results in various errors, such as the onUrl
module passing something along that doesn't parse with the other method,
or detection of bogus urls.

For a specific example:

'!g matthew 5:29' parses as valid with 'url.parse', but throws an error
with new URL, so before `onUrl` would pass that along as a url with bad
results.

using `url.parse` everywhere could solve this, but `new URL` is the
newer api, so I'd rather standardize on that.